requirements.txt
Django
djangorestframework
rest_framework_simplejwt
tqdm
requests
